requirements.txt
DBUtils==3.0.2
Flask==2.1.2
Flask_Cors==3.0.10
Flask_SocketIO==4.3.0
matplotlib==3.5.2
numpy==1.21.6
opencv_python==4.6.0.66
Pillow==9.2.0
prettytable==3.3.0
PyMySQL==1.0.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
